1768: update solution

+++ algorithm_1000-1999/1768-Merge Strings Alternately/main.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     string mergeAlternately(string word1, string word2) {
-        size_t n = std::min(word1.size(), word2.size());
+        int n = std::min(word1.size(), word2.size());
 
         string ret;
         for (int i = 0; i < n; ++i) {
